+++ src/presentation/controllers/download-subtitle/download-subtitle-protocols.ts
@@ -1,0 +1,3 @@
+export { GetSubtitle } from '../../../domain/usecases/get-subtitle'
+export { SubtitleModel } from '../../../domain/models/subtitle'
+export { Controller, HttpRequest, HttpResponse } from '../../protocols'
